+++ src/Manager/FieldPaletteModelManager.php
@@ -13,6 +13,11 @@ use HeimrichHannot\FieldpaletteBundle\Model\FieldPaletteModel;
 class FieldPaletteModelManager
 {
     /**
+     * @var FieldPaletteModel
+     */
+    protected $modelInstance = null;
+
+    /**
      * Return a new model instance.
      *
      * @return FieldPaletteModel
@@ -25,6 +30,23 @@ class FieldPaletteModelManager
     }
 
     /**
+     * Returns an FieldPaletteModel instance for model calls without creating a new one (only if no already instantiated or has modified table).
+     *
+     * @return FieldPaletteModel
+     */
+    public function getInstance()
+    {
+        if (!$this->modelInstance) {
+            $this->modelInstance = $this->createModel();
+        } elseif (!$this->modelInstance->hasTable()) {
+            unset($this->modelInstance);
+            $this->modelInstance = $this->createModel();
+        }
+
+        return $this->modelInstance;
+    }
+
+    /**
      * Returns a new model instance with given table set.
      *
      * @param string $table
@@ -34,7 +56,7 @@ class FieldPaletteModelManager
      */
     public function createModelByTable(string $table)
     {
-        $model = new FieldPaletteModel();
+        $model = $this->createModel();
         if (!empty($table)) {
             $model->setTable($table);
         }